Add hotfix for container issue and landing page logos duplication.
