+++ resources/ut-tei/src/scss/header/custom/polk/_polk-hero.scss
@@ -70,7 +70,7 @@ html {
   width: 100%;
   height: 100%;
   background-repeat: no-repeat;
-  background-position: 100% 45%;
+  background-position: 75% 45%;
   background-size: auto 375px;
   background-color: $polk;
 }
