+++ resources/ut-tei/src/scss/header/custom/polk/_polk-landing-page.scss
@@ -1,4 +1,4 @@
-.polk-logos {
+.content .polk-logos {
   margin-top: 123px;
   margin-bottom: 76px;
