+++ search.html
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <div data-template="templates:surround" data-template-with="templates/page.html" data-template-at="content">
     <div data-template="templates:include" data-template-path="templates/menu.html"/>
-    <section class="container-fluid">
+    <section class="container">
         <main xmlns:i18n="http://exist-db.org/xquery/i18n" class="row search">
             <div class="col-md-4 col-md-push-8">
                 <div class="search-help">
