+++ templates/menu.html
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <section class="utk_header">
     <nav xmlns:i18n="http://exist-db.org/xquery/i18n" class="navbar navbar-default" role="navigation" data-template="i18n:translate" data-catalogues="resources/i18n">
-        <div class="container-fluid" data-template="browse:fix-links">
+        <div class="container" data-template="browse:fix-links">
             <div class="navbar-collapse collapse" id="navbar-collapse-1">
                 <ul class="nav navbar-nav">
                     <li class="dropdown" data-template="browse:show-if-logged-in">
