+++ view.html
@@ -24,7 +24,7 @@
                             <!-- Document title -->
                             <h5 class="tp-column" data-template="pages:navigation-title"/>
                         </section>
-                        <section id="document-wrapper" class="tp-row">
+                        <section id="document-wrapper" class="tp-row container">
                             <!-- Document pane -->
                             <div id="document-pane" class="tp-column">
                                 <div data-template="pages:view" class="content"/>
